Use Qt data types instead of native types
Probably an overkill, but it may help us to ensure that everything works
as expected on every platform

+++ include/LibDS/Core/Common.h
@@ -73,12 +73,12 @@ enum DS_CommunicationStatus {
  * @brief The DS_Joystick struct
  */
 struct LIB_DS_DECL DS_Joystick {
-    int numAxes;
-    int numButtons;
-    int numPovHats;
+    quint8 numAxes;
+    quint8 numButtons;
+    quint8 numPovHats;
 
     bool* buttons;
-    int* povHats;
+    quint8* povHats;
     double* axes;
 };
 
@@ -97,7 +97,7 @@ QString LIB_DS_DECL DS_GetTimezoneCode();
  *     - \c DS_GetStaticIp( 118, 3) would return \c 10.01.18.3
  *     - And so on...
  */
-QString LIB_DS_DECL DS_GetStaticIp (int team, int host);
+QString LIB_DS_DECL DS_GetStaticIp (quint16 team, quint16 host);
 
 /**
  * Returns an user-friendly string given the inputed robot control mode
